backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.0
pdfminer.six==20231228
tabula-py==2.8.2
camelot-py==0.10.1
pandas==2.1.4
numpy==1.24.3
spacy==3.7.2
transformers==4.35.2
sentence-transformers==2.2.2
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
faiss-cpu==1.7.4
neo4j==5.14.1
redis==5.0.1
networkx==3.2.1
scikit-learn==1.3.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
celery==5.3.4
rdflib==7.0.0
geopandas==0.14.1
folium==0.15.0
plotly==5.17.0
streamlit==1.28.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0